requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
django-filters==0.2.1
djangorestframework==3.9.1
djangorestframework-jsonapi==2.7.0
inflection==0.3.1
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
six==1.12.0
